backend/app/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
aiohttp==3.9.1
asyncio==3.4.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
Jinja2==3.1.2
Werkzeug==2.3.7
SQLAlchemy==2.0.23